perl-module: Use Autotest for testing Perl.
The output of the existing Perl test is difficult to understand, in my
opinion.  When everything goes right, it prints out plenty of stuff.
When something goes wrong, it prints more stuff, but that stuff usually
doesn't make it clear to me what went wrong or where.

Moving it into Autotest helps in two ways.  When everything goes right,
all you get is a clear summary that says that.  When something goes wrong,
you get clear log files that explain where.

(This commit actually duplicates the Perl module tests: it makes Autotest
copies of each of the individual tests, and Autotest also runs the
t/Pspp.t unit test.  The former are easier to understand; the latter is
what Perl hackers apparently expect to see.)