Document system file record type 7, subtype 16 as 64-bit number of cases.
I found this record type and subtype in all .sav files written by SPSS 14
and later (and in one file written by SPSS 13).  After scratching my head
a bit I realized that it always contained the same value as the "ncases"
field in the top-level file header, but written as a 64-bit number.  So
presumably the purpose is to allow for a 64-bit count of cases.