Convert syntax files to UTF-8 before loading them.
Convert the contents of a syntax file to UTF-8 before loading
it into the syntax window. Fixes bug #29733
This change assumes that the files to be loaded are in the
encoding of the current locale.  If this assumption is not
true, then an error will be displayed and the file not loaded.