tmpfile: Rename "struct tmpfile" to "struct temp_file".
Mingw builds were failing because Gnulib was replacing "tmpfile" by
"rpl_tmpfile", and this was visible only to some users of "struct tmpfile".
The safest fix seems to be just renaming struct tmpfile, so that's what
this commit does.

Thanks to Harry Thijssen <pspp@sjpaes.nl> for reporting the problem.

Bug #29965.