Do not treat isolated CR in input data as new-line.
User "tvw" on IRC reported that PSPP failed to parse data that SPSS 18
accepted.  We found that the problem was carriage return (CR) characters
in the middle of a line.  PSPP treated these as new-lines, but SPSS did
not.  This commit adopts the SPSS behavior in PSPP and adjusts one test
that checks this behavior.

This will break reading some old Mac OS files, since Mac OS before version
10 used CR without LF as new-line.  Time will tell whether this is a real
problem for our users.