cairo: Remove support for printed headers.
The printed headers on each page do not very attractive, and the
information that they contain by default is not very useful.  Because an
upcoming commit makes them harder to properly implement, simply remove
them for now.

With this commit, the cairo driver no longer prints titles and subtitles
at all.  This needs to be fixed.