Revert unintended change to tests_libpspp_str_test_LDADD.
When I was preparing the previous commit I tested a change to
tests_libpspp_str_test_LDADD and then (I thought) reverted it.  But my
revert actually changed the order.  Oops.

John Darrington reports that the new order causes link problems for him, so
I'm putting the order back to the original, with this commit.