cairo: Fix handling of font sizes.
Until now this code used pango_font_description_set_absolute_size() to set
the font size.  This sets the font size in device units, e.g. pixels on a
display.  But font sizes are actual specified by the user in points, so
we should instead use pango_font_description_get_size().  This commit
makes that change.

That change is simple, but it necessitates a larger change.
Until now the cairo driver has just assumed that the size of a font is
the size that we assigned it.  But this is no longer the case, since
points and pixels are now different.  So now we measure the actual width
and height of characters and use those as the font size.  At the same
time, the code no longer ignores font sizes included as part of font
description strings, and allows different fonts to have different sizes.

With plenty of help from John Darrington.