lexer: Get rid of lex_negative_to_dash().
This function stands in the way of replacing the lexer in a
straightforward way, because it disrupts the stream of tokens: the
lexer can't tell in advance whether, e.g., "-4" is one token or
two.  Fortunately, it is not used in very many places and is
relatively easy to remove.