AGGREGATE: New file aggregate.h containing the summary function specifications
Added a header file containing the declaration and type of the agr_func_tab
table, and expanded that type to contain a description of the function and
an explicit declaration of whether a source variable is permitted/needed.
This change is primarily for the benefit of the forthcoming dialog box for
the aggregate command.  However it also (in my opinion) makes the existing
implementation a little easier to understand.