Flush the journal after every write.
Unexpected bugs in the parser, engine or in procedures can cause crashes
in the GUI.  If the journal is not flushed *before* running the syntax, then
no record of that syntax exists, and the bug can be hard to reproduce.
This change ensures that the syntax is written to the journal before it is
processed.