Categoricals.c Create entries for all groups.
Although for the purposes of creating a covariance matrix, only
n - 1 entries are needed,  for other purposes (such as descriptive
statistics, contrasts, planned comparisons etc) all n entries are required.
(where n is the number of distinct values).  This change updates
categories.c so that all n categories are generated.  Covariance matrix
routines are free to ignore the ones not of interest.