lexer: Improve translatability of lex_error().
In general I believe that it is easier to translate full sentences or
clauses.  It seems likely that "Syntax error %s at %s." where the first
%s is an arbitrary English phrase would be impossible to translate
grammatically into some languages, so this changes the general form to
"Syntax error at %s: %s." which seems more likely to be translatable while
not changing the English meaning.