message: Consistently initialize locator; use 0 for "no line number".
A few of the callers of msg_emit() did not initialize the "where" member
of the struct msg, because they expect that msg_emit() will do it for them.
This is currently correct, but I intend to soon introduce the ability for
msg_emit()'s caller to specify a location.  With that change, it will be
important for the caller to initialize this member, so this commit makes
sure of that.

At the same time, some callers were using -1 as the default value that
means "no line number" and others were using 0.  This commit standardizes
on the latter.