syntax-file: Open syntax files immediately, instead of delaying.
Until now, syntax files have been opened only when the first line needs to
be read.  This commit changes behavior to open syntax files as soon as the
syntax file source is created.

This fixes the behavior INSERT CD=YES when a relative file path is
specified.  Without this commit, INSERT checks that the file exists, then
it creates the syntax file source, then it chdirs into the syntax file's
directory, and then the syntax file source tries to open the file, which
fails because the relative path is no longer correct.  With this commit,
the file gets opened before the chdir, so it succeeds.

The following commit will add a test to prevent regression.