tests: Distribute testsuite.at and loosen its dependencies.
testsuite.at has to be distributed or "make" from a distributed tarball
will have to make it, which means that the system used for building needs
Autoconf installed and will need to write to the source directory.

testsuite.at cannot depend on the Makefile or it will have to be rebuilt
every time "configure" runs, which would also cause the problems mentioned
above.