doc: Always build ni.texi, tut.texi in source directory.
When the source and build directories are different, ni.texi and tut.texi
were being built in the build directory.  This meant that, since
$(builddir)/doc is not in the TeX search path, but $(builddir) is (because
it is the working directory when TeX is invoked), the @include commands
that include ni.texi and tut.texi had to specify doc/ explicitly.
However, doing that broke the case where the source and build directories
were different but ni.texi or tut.texi were in the source directory (which
happens when building from a tarball or, similarly, running "make
distcheck"), because $(top_srcdir) is not in the TeX search path (only
$(top_srcdir)/doc is).

The easiest solution appears to be to always make sure that ni.texi and
tut.texi are in the source directory.  Then they are always in the TeX
search path (as long as we drop the doc/ prefix) and everything should
work out OK.

Reported by Michel Boaventura <michel@michelboaventura.com>.