render: Fix cells spanning otherwise completely empty rows/columns.
When adjacent rows or columns contain only spanned cells and no rules give
the rows or columns a natural minimum height or width, this triggers a
special case in distribute_spanned_width().  This special case had not been
tested and (therefore) was broken.  This commit fixes the problem and adds
test cases to prevent it from recurring.

Bug #31346.