csv: Better support separators other than comma or tab.
Before this commit, although any separator could be specified for cvs
output, the set of characters that triggered quoting was fixed, and only
really suitable for comma and tab separators.  This commit fixes the
problem, making sure that whenever the specified separator appears in a
field, that field will be quoted.

Problem noticed during a discussion with Björn Manke <bjoernmanke@gmx.net>
on pspp-users.