DEBUG XFORM FAIL: Issue an error message as part of transformation.
PSPP only returns a failure exit status if an error message is issued, but
DEBUG XFORM FAIL doesn't issue an error message.  This means that, if PSPP
has no other errors, then it's difficult to tell that the transformation
really did what it should have.  This commit fixes the potential problem.