MODIFY VARIABLES: Eliminate VAR_NAME_LEN limit on variable names.
This is actually a general code improvement, since it eliminates memory
allocation and copying that was not actually necessary.

Most uses of VAR_NAME_LEN within PSPP are wrong due to encoding issues:
the limit applies to variable names in the encoding used by the data
set, but most uses of VAR_NAME_LEN actually limit the length of a name
in UTF-8.  The UTF-8 representation of a name can be longer or shorter
than its representation in the data set encoding, so it seems best to
eliminate references to VAR_NAME_LEN entirely.