variable-parser: Drop VAR_NAME_LEN restriction from var_set_lookup_var_idx().
This restriction is purely artificial, as part of an assertion.  Since
longer variable names are going to have to be supported, remove it.

Most uses of VAR_NAME_LEN within PSPP are wrong due to encoding issues:
the limit applies to variable names in the encoding used by the data
set, but most uses of VAR_NAME_LEN actually limit the length of a name
in UTF-8.  The UTF-8 representation of a name can be longer or shorter
than its representation in the data set encoding, so it seems best to
eliminate references to VAR_NAME_LEN entirely.