FLIP: Eliminate false dependency on VAR_NAME_LEN.
The size of this buffer really shouldn't have anything to do with
VAR_NAME_LEN but with the string about to be put into it.

Most uses of VAR_NAME_LEN within PSPP are wrong due to encoding issues:
the limit applies to variable names in the encoding used by the data
set, but most uses of VAR_NAME_LEN actually limit the length of a name
in UTF-8.  The UTF-8 representation of a name can be longer or shorter
than its representation in the data set encoding, so it seems best to
eliminate references to VAR_NAME_LEN entirely.