Change terminology from "active file" to "active dataset".
I plan to introduce the concept of multiple datasets.  One of these is
active at any given time, and the others are inactive.  Each one is
similar to what has been called the "active file" until now.  Thus, it
is natural to rename the "active file" to the "active dataset".  I
guess that this greater uniformity of terminology will cause less
user confusion.