Fix up build following dropping ulc-width-linebreaks module.
Commit b5cebf00d5e "Smake: Remove  module"
unexpectedly broke the build, because ulc-width-linebreaks had several
indirect dependencies that PSPP actually used but were not in the
list of Gnulib modules in Smake.  This fixes the problem.

The oddest consequence fixed by this commit has to do with
build-aux/config.rpath.  AM_GNU_GETTEXT requires this file, which is
supplied by only a few Gnulib modules: havelib, gettext, and
threadlib.  ulc-width-linebreaks depended indirectly on havelib, but
PSPP did not otherwise depend on any of these modules.  It seemed that
the best fix was simply to use the Gnulib gettext module, which is
what this commit does.