gui: Fix crash in Paired T-Test dialog on selecting first variable.
When a variable is moved into the list of selected variables in the
Paired T-Test dialog, that row of the treeview has one nonnull
variable and one null variable.  Calling var_get_name(NULL) causes a
segfault.

I'm not certain that this is the correct fix, but it fixes the
segfault.  The missing variable is now shown as <unset> until a second
variable is moved into the treeview.

Bug #32958.