Update code that assumed a dictionary's encoding could be NULL.
Until recently, I think, it was possible for dict_get_encoding()
to return NULL.  It is no longer possible as far as I can see,
so this commit adds a comment that says so to the definition
of dict_get_encoding().  It also updates a few callers that
had checks for null pointers.