sys-file-writer: Omit empty multiple response sets records.
Multiple response sets are divided into two records in a .sav
file: one record with subtype 7 and one with subtype 19.  PSPP
was incorrectly writing both records even if one of them was
empty.  This commit makes it omit the empty record.

Thanks to Curt Reinhold <Curt.Reinhold@psychonomics.de> for
providing the .sav file that led to this discovery.