sys-file-reader: Reduce invalid format from error to warning.
gaby jongenelen <geebie21@gmail.com> provided a .sav file
written by "MS Windows Release 13.0 spssio32.dll" that had
zeros in the "write" format field for several numeric
variables.  Until now, PSPP treated this as an error and
rejected the file.  This commit changes the error to a warning
and suppresses the diagnostic entirely for this case of a
zero value.