New widget PsppireValChooser.
This widget provides a user interface allowing the user to input the
a value, or range of values.  This was previously done "by hand" in
the recode dialog, but was awkward to maintain, and clumsy.  Upcoming
dialogs will need something very similar. So now seemed like a good
time to encapsulate it.