gui: New type PsppireEmptyListStore.
PsppireEmptyListStore is an GtkTreeModel implementation that has a
client-specified number of rows and zero columns.  It is a useful model for
GtkTreeView or PsppSheetView when the client can easily synthesize cell
data using a callback set with gtk_tree_view_column_set_cell_data_func().
In that situation, GtkListStore can be wasteful (because it uses a lot of
memory to store what does not need to be stored) and situation-specific
custom models require additional boilerplate.

This commit also modified the text data import dialog code to use
PsppireEmptyListStore.  Future commits will add other users.