range-set: Skip some tests on llvm-gcc 4.2.
These tests seem to trigger a bug in llvm-gcc 4.2 on Mac OS X
10.8.0.  Experimentation shows that the bug is more likely to
be triggered by the test code than the code under test, so skip
them.

I spent some time trying to narrow down the problem enough to
make a bug report to the llvm-gcc developers, but I ran out of
time before I finished with that.

Thanks to Jeremy Lavergne <jeremy@lavergne.gotdns.org> for
reporting the problem.