gui: Streamline pspp.desktop
Remove the extension for Icon (unnecessary, in case in
future we want to provide for example SVG icon) and removed
unnecessary keys. Do we know for fact that we will never ever
support Notifications? And I guess if we provide the .desktop and
executable as part of the same tarball/package there is not much
chance we would have binary missing. I have also added a Czech
translation of keys.

Signed-off-by: Matìj Cepl <mcepl@redhat.com>